split socket.go and dialer.go
